requirements.txt
flask-restplus==0.9.2
PyYAML==3.12
Flask-SQLAlchemy==2.1